GC_Function/requirements.txt
google-cloud-pubsub == 2.19.0
google-api-core == 2.15.*